DS-1064: Move IAM related classes into iam package
